+++ isitfit/cli.py
@@ -8,9 +8,8 @@ import logging
 logger = logging.getLogger('isitfit')
 
 from .mainManager import MainManager
-
-
 import click
+from tabulate import tabulate
 
 @click.command()
 @click.option('--debug', is_flag=True)
@@ -26,17 +25,28 @@ def cli(debug):
     mm = MainManager()
     n_ec2, sum_capacity, sum_used, cwau = mm.get_ifi()
     logger.info("... done")
+    
+    dt_start = mm.StartTime.strftime("%Y-%m-%d")
+    dt_end   = mm.EndTime.strftime("%Y-%m-%d")
+    
+    table = [
+      ["Analysis start date", "%s"%dt_start],
+      ["Analysis end date", "%s"%dt_end],
+      ["Number of analysed EC2 machines", "%i"%n_ec2],
+      ["Billed cost per hour", "%0.2f $/hour"%sum_capacity],
+      ["Used cost per hour", "%0.2f $/hour"%sum_used],
+      ["CWAU = Used / Billed * 100", "%0.0f %%"%cwau],
+    ]
+    
     logger.info("")
     logger.info("Summary:")
-    logger.info("Number of EC2 machines = %i"%n_ec2)
-    logger.info("Billed cost per hour = %0.2f $/hour"%sum_capacity)
-    logger.info("Used cost per hour = %0.2f $/hour"%sum_used)
-    logger.info("CWAU = Used / Billed * 100 = %0.0f %%"%cwau)
+    logger.info("")
+    logger.info(tabulate(table, headers=['Field', 'Value']))
     logger.info("")
     logger.info("For reference:")
     logger.info("* CWAU >= 70% is well optimized")
     logger.info("* CWAU <= 30% is underused")
-    logger.info("* CWAU in isitfit version 0.1.2 is based on CPU utilization only (and not memory utilization)")
+    logger.info("* isitfit version 0.1 is based on CPU utilization only (and not yet on memory utilization)")
 
 
 if __name__ == '__main__':